backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
supabase>=2.3.0
google-generativeai>=0.4.0
python-doctr[torch]>=0.8.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
